refactor: merge patch.go into user_repository.go
